Separating GEM-specific functionality from reg_interface
